Sprint 5

+++ shared/src/persistence/dynamo/workitems/syncWorkItems.js
@@ -8,19 +8,40 @@ exports.syncWorkItems = async ({
   caseToSave,
   currentCaseState,
 }) => {
-  for (let workItem of caseToSave.workItems || []) {
-    const existing = (currentCaseState.workItems || []).find(
-      i => i.id === workItem.id,
+  let currentWorkItems = [];
+  let newWorkItems = [];
+  (currentCaseState.documents || []).forEach(
+    document =>
+      (currentWorkItems = [...currentWorkItems, ...(document.workItems || [])]),
+  );
+  (caseToSave.documents || []).forEach(
+    document =>
+      (newWorkItems = [...newWorkItems, ...(document.workItems || [])]),
+  );
+
+  for (let workItem of newWorkItems) {
+    const existing = currentWorkItems.find(
+      item => item.workItemId === workItem.workItemId,
     );
     if (!existing) {
+      if (workItem.assigneeId) {
+        await persistence.createMappingRecord({
+          pkId: workItem.assigneeId,
+          skId: workItem.workItemId,
+          type: 'workItem',
+          applicationContext,
+        });
+      }
+
       await persistence.createMappingRecord({
-        pkId: workItem.assigneeId,
+        applicationContext,
+        pkId: workItem.section,
         skId: workItem.workItemId,
         type: 'workItem',
-        applicationContext,
       });
 
       await client.put({
+        applicationContext,
         TableName: `efcms-${applicationContext.environment.stage}`,
         Item: {
           pk: workItem.workItemId,
@@ -33,9 +54,9 @@ exports.syncWorkItems = async ({
       if (workItem.assigneeId !== existing.assigneeId) {
         // the item has changed assignees, delete item
         await exports.reassignWorkItem({
+          applicationContext,
           existingWorkItem: existing,
           workItemToSave: workItem,
-          applicationContext,
         });
       }
     }
@@ -43,21 +64,39 @@ exports.syncWorkItems = async ({
 };
 
 exports.reassignWorkItem = async ({
+  applicationContext,
   existingWorkItem,
   workItemToSave,
-  applicationContext,
 }) => {
-  await persistence.deleteMappingRecord({
-    pkId: existingWorkItem.assigneeId,
-    skId: workItemToSave.workItemId,
-    type: 'workItem',
-    applicationContext,
-  });
+  if (existingWorkItem.assigneeId) {
+    await persistence.deleteMappingRecord({
+      applicationContext,
+      pkId: existingWorkItem.assigneeId,
+      skId: workItemToSave.workItemId,
+      type: 'workItem',
+    });
+  }
+
+  if (existingWorkItem.section !== workItemToSave.section) {
+    await persistence.deleteMappingRecord({
+      applicationContext,
+      pkId: existingWorkItem.section,
+      skId: existingWorkItem.workItemId,
+      type: 'workItem',
+    });
+
+    await persistence.createMappingRecord({
+      applicationContext,
+      pkId: workItemToSave.section,
+      skId: workItemToSave.workItemId,
+      type: 'workItem',
+    });
+  }
 
   await persistence.createMappingRecord({
+    applicationContext,
     pkId: workItemToSave.assigneeId,
     skId: workItemToSave.workItemId,
     type: 'workItem',
-    applicationContext,
   });
 };
